Add CMakeLists and Kconfig
The Kconfig file allows the user to choose which features
of the modem lib to enable.
The CMakeLists file converst the Kconfig settings to defines
used in the modem library and includes the required sources and headers.

Only the minimum amount of headers are exposed outside of the library,
so that the insides of the library will not be called by users on
accident.